NCB collection code, client and server.
This is the BlueJ client and Python server for collecting novice
compilation behavior data from BlueJ. The server is potentially more
general purpose (and needs serious consideration with respect to
security).

This code has been available elsewhere on the net in the past, but is
being moved here as it is easier to use existing infrastructure than to
continue maintaining my own for serving this code to the world.